psh: modularization of psh applets
 - each applet separated into its own directory
 - each applet registers its command(s) in list using constructors
 - interactive psh separated as applet "pshapp"
 - login separated as applet "login"
 - remodelled "help" to work with modularized applets

Applets modularization into directories does not support subtrees in those directories.
Each applet must register itself using staticly defined list element passed to list in psh.
Each applet can register more than one command in psh but no ordering is done during registering.
A command may be hidden from "help" by not specifying info function (passing NULL in constructor).

Above changes achieve good psh modularization and simplifies process of new applet creation.

JIRA: PD-74
